+++ accounting/DataBase/DTOs/LoansDTO.cs
@@ -8,12 +8,12 @@ namespace accounting.DataBase.DTOs
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Key]
-        public int Id { get; set; }
+        public ushort Id { get; set; }
 
-        public int Amount { get; set; }
-        public int InstallmentsCount { get; set; }
+        public ulong Amount { get; set; }
+        public byte InstallmentsCount { get; set; }
         public string LendDate { get; set; } = null!;
-        public int AccountId { get; set; }
+        public ushort AccountId { get; set; }
         public string? PersonalAccountNumber { get; set; }
         [ForeignKey("AccountId")] public AccountDTO Account { get; set; }
         public List<LoanInstallmentsDTO>? Installments { get; set; }
